Refactor the Entity class.
As it wasn't a normal case class, circe can't encode it easily.
I've swapped this out for a normal case class.
