+++ src/main/scala/uk/gov/nationalarchives/dp/client/EntityClient.scala
@@ -12,6 +12,7 @@ import java.time.format.DateTimeFormatter
 import java.util.UUID
 import scala.xml.{Elem, XML}
 import uk.gov.nationalarchives.dp.client.Client._
+import uk.gov.nationalarchives.dp.client.Entities.Entity
 
 trait EntityClient[F[_], S] {
   def metadataForEntity(entity: Entity, secretName: String): F[Seq[Elem]]
@@ -74,9 +75,9 @@ object EntityClient {
           .map(url => getApiResponseXml(url, token))
           .sequence
         allUrls <- dataProcessor.allBitstreamUrls(allGenerationElements)
-        bitStreams <- allUrls.map(url => getApiResponseXml(url, token)).sequence
-        allBitstreams <- dataProcessor.allBitstreamInfo(bitStreams)
-      } yield allBitstreams
+        bitstreamXmls <- allUrls.map(url => getApiResponseXml(url, token)).sequence
+        allBitstreamInfo <- dataProcessor.allBitstreamInfo(bitstreamXmls)
+      } yield allBitstreamInfo
     }
 
     override def metadataForEntity(entity: Entity, secretName: String): F[Seq[Elem]] = {
